Dev koputo

+++ TypeThread.cpp
@@ -38,7 +38,7 @@ uint32_t TypeThread::getAllThreads(TypeThread **dstPtr) {
                     strncpy(tThread[r].title, row[5], flen);
                     memset(tThread[r].device_uuid, 0, 37);
                     strncpy(tThread[r].device_uuid, row[2], 36);
-                    tThread[r].speed = static_cast<uint16_t>(strtol(row[4], nullptr, 10));
+                    tThread[r].speed = static_cast<uint64_t>(strtol(row[4], nullptr, 10));
                     tThread[r].status = strtol(row[6], nullptr, 10);
                     tThread[r].work = strtol(row[7], nullptr, 10);
 //                    tThread[r].deviceType = strtol(row[6], nullptr, 10);
