+++ TypeThread.h
@@ -15,7 +15,7 @@ public:
     char port[15];
     char device_thread[15];
     char device_uuid[37];
-    uint16_t speed;
+    uint64_t speed;
 //    uint16_t deviceType;
     char deviceType[37]; // uuid
     std::time_t lastDate;
